I moved the structure files to a new folder to structure
