+++ src/index.js
@@ -1,12 +1,11 @@
 import React from "react";
 
 import { WrappLetter } from "./process/wrappLetter.process";
-import { error__main_filter__ } from "./error";
 import { error__props_filter__ } from "./error/props-filter.error";
 
 import { memo__process } from "./process/memo.process";
 import { __Text__process__ } from "./process/text.class";
-import { __Structure__process__ } from "./process/structure.class";
+import { __Structure__process__ } from "./process/structure/structure.class";
 import { __TextOptions__process__ } from "./process/textOptions/textOptions.class";
 
 function WL(
